api/requirements.txt
## The following requirements were added by pip freeze:
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2